Update for critical security patches.
git-svn-id: https://plugins.svn.wordpress.org/wp-easy-gallery/trunk@669527 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ admin/add-images.php
@@ -11,44 +11,52 @@ $galleryResults = $wpdb->get_results( "SELECT * FROM $easy_gallery_table" );
 
 //Select gallery
 if(isset($_POST['select_gallery']) || isset($_POST['galleryId'])) {
-	$gid = (isset($_POST['select_gallery'])) ? $_POST['select_gallery'] : $_POST['galleryId'];
-	$imageResults = $wpdb->get_results( "SELECT * FROM $easy_gallery_image_table WHERE gid = $gid ORDER BY sortOrder ASC" );
-	$gallery = $wpdb->get_row( "SELECT * FROM $easy_gallery_table WHERE Id = $gid" );
+	if(check_admin_referer('wpeg_select_gallery','wpeg_select_gallery')) {
+	  $gid = intval((isset($_POST['select_gallery'])) ? mysql_real_escape_string($_POST['select_gallery']) : mysql_real_escape_string($_POST['galleryId']));
+	  $imageResults = $wpdb->get_results( "SELECT * FROM $easy_gallery_image_table WHERE gid = $gid ORDER BY sortOrder ASC" );
+	  $gallery = $wpdb->get_row( "SELECT * FROM $easy_gallery_table WHERE Id = $gid" );
+	}
 }
 
 //Add image
 if(isset($_POST['galleryId']) && !isset($_POST['switch'])) {
-	$gid = $_POST['galleryId'];
-	$imagePath = $_POST['upload_image'];
-	$imageTitle = $_POST['image_title'];
-	$imageDescription = $_POST['image_description'];
-	$sortOrder = 0 + $_POST['image_sortOrder'];
-	$imageAdded = $wpdb->insert( $easy_gallery_image_table, array( 'gid' => $gid, 'imagePath' => $imagePath, 'title' => $imageTitle, 'description' => $imageDescription, 'sortOrder' => $sortOrder ) );
-	
-	if($imageAdded) {
-	?>
-		<div class="updated"><p><strong><?php _e('Image saved.' ); ?></strong></p></div>  
-	<?php }
-	//Reload images
-	$imageResults = $wpdb->get_results( "SELECT * FROM $easy_gallery_image_table WHERE gid = $gid ORDER BY sortOrder ASC" );
+	if(check_admin_referer('wpeg_insert_image','wpeg_insert_image')) {
+	  $gid = intval($_POST['galleryId']);
+	  $imagePath = $_POST['upload_image'];
+	  $imageTitle = $_POST['image_title'];
+	  $imageDescription = $_POST['image_description'];
+	  $sortOrder = intval($_POST['image_sortOrder']);
+	  $imageAdded = $wpdb->insert( $easy_gallery_image_table, array( 'gid' => $gid, 'imagePath' => $imagePath, 'title' => $imageTitle, 'description' => $imageDescription, 'sortOrder' => $sortOrder ) );
+	  
+	  if($imageAdded) {
+	  ?>
+		  <div class="updated"><p><strong><?php _e('Image saved.' ); ?></strong></p></div>  
+	  <?php }
+	  //Reload images
+	  $imageResults = $wpdb->get_results( "SELECT * FROM $easy_gallery_image_table WHERE gid = $gid ORDER BY sortOrder ASC" );
+	}
 }
 
 //Edit image
 if(isset($_POST['edit_image'])) {
-	$imageEdited = $wpdb->update( $easy_gallery_image_table, array( 'imagePath' => $_POST['edit_imagePath'], 'title' => $_POST['edit_imageTitle'], 'description' => $_POST['edit_imageDescription'], 'sortOrder' => $_POST['edit_imageSort'] ), array( 'Id' => $_POST['edit_image'] ) );
-		
-		?>  
-        <div class="updated"><p><strong><?php _e('Image has been edited.' ); ?></strong></p></div>  
-        <?php
+	if(check_admin_referer('wpeg_edit_image','wpeg_edit_image')) {
+		$imageEdited = $wpdb->update( $easy_gallery_image_table, array( 'imagePath' => $_POST['edit_imagePath'], 'title' => $_POST['edit_imageTitle'], 'description' => $_POST['edit_imageDescription'], 'sortOrder' => $_POST['edit_imageSort'] ), array( 'Id' => intval($_POST['edit_image']) ) );
+			
+			?>  
+			<div class="updated"><p><strong><?php _e('Image has been edited.' ); ?></strong></p></div>  
+			<?php
+	}
 }
 
 // Delete image
 if(isset($_POST['delete_image'])) {
-	$wpdb->query( "DELETE FROM $easy_gallery_image_table WHERE Id = '".$_POST['delete_image']."'" );
+	if(check_admin_referer('wpeg_delete_image','wpeg_delete_image')) {
+		$wpdb->query( "DELETE FROM $easy_gallery_image_table WHERE Id = '".$_POST['delete_image']."'" );
 		
 		?>  
         <div class="updated"><p><strong><?php _e('Image has been deleted.' ); ?></strong></p></div>  
-        <?php	
+        <?php
+	}
 }
 
 ?>
@@ -59,18 +67,19 @@ if(isset($_POST['delete_image'])) {
 	<?php if(!isset($_POST['select_gallery']) && !isset($_POST['galleryId'])) { ?>
     <p>Select a galley</p>		
     <form name="gallery" method="post" action="<?php echo str_replace( '%7E', '~', $_SERVER['REQUEST_URI']); ?>">
-    	<select name="select_gallery" onchange="gallery.submit()">
+    	<?php wp_nonce_field('wpeg_select_gallery','wpeg_select_gallery'); ?>
+        <select name="select_gallery" onchange="gallery.submit()">
         	<option> - SELECT A GALLERY - </option>
 			<?php
 				foreach($galleryResults as $gallery) {
-					?><option value="<?php echo $gallery->Id; ?>"><?php echo $gallery->name; ?></option>
+					?><option value="<?php _e($gallery->Id); ?>"><?php _e($gallery->name); ?></option>
                 <?php
 				}
 			?>
         </select>
     </form>
     <?php } else if(isset($_POST['select_gallery']) || isset($_POST['galleryId'])) { ?>    
-    <h3>Gallery: <?php echo $gallery->name; ?></h3>
+    <h3>Gallery: <?php _e($gallery->name); ?></h3>
     <form name="switch_gallery" method="post" action="<?php echo str_replace( '%7E', '~', $_SERVER['REQUEST_URI']); ?>">
     <input type="hidden" name="switch" value="true" />
     <p style="float: left;"><input type="submit" name="Submit" class="button-primary" value="Switch Gallery" /></p>
@@ -79,7 +88,8 @@ if(isset($_POST['delete_image'])) {
     <div style="Clear: both;"></div>
     
     <form name="add_image_form" action="<?php echo str_replace( '%7E', '~', $_SERVER['REQUEST_URI']); ?>" method="post">
-    <input type="hidden" name="galleryId" value="<?php echo $gallery->Id; ?>" />
+    <input type="hidden" name="galleryId" value="<?php _e($gallery->Id); ?>" />
+    <?php wp_nonce_field('wpeg_insert_image','wpeg_insert_image'); ?>
     <table class="widefat post fixed eg-table">
     	<thead>
         <tr>
@@ -136,19 +146,21 @@ if(isset($_POST['delete_image'])) {
         <tbody>        	
         	<?php foreach($imageResults as $image) { ?>				
             <tr>
-            	<td><img src="<?php echo $image->imagePath; ?>" width="75" alt="" /></td>
+            	<td><img src="<?php _e($image->imagePath); ?>" width="75" alt="" /></td>
                 <td>
                 	<form name="edit_image_form" action="<?php echo str_replace( '%7E', '~', $_SERVER['REQUEST_URI']); ?>" method="post">
-                	<input type="hidden" name="edit_image" value="<?php echo $image->Id; ?>" />
-                	<p><strong>Image Path:</strong> <input type="text" name="edit_imagePath" size="75" value="<?php echo $image->imagePath; ?>" /></p>
-                    <p><strong>Image Title:</strong> <input type="text" name="edit_imageTitle" size="20" value="<?php echo $image->title; ?>" /></p>
-                    <p><strong>Image Description:</strong> <input type="text" name="edit_imageDescription" size="75" value="<?php echo $image->description; ?>" /></p>
-                    <p><strong>Sort Order:</strong> <input type="text" name="edit_imageSort" size="10" value="<?php echo $image->sortOrder; ?>" /></p>
+                	<input type="hidden" name="edit_image" value="<?php _e($image->Id); ?>" />
+                    <?php wp_nonce_field('wpeg_edit_image', 'wpeg_edit_image'); ?>                    
+                	<p><strong>Image Path:</strong> <input type="text" name="edit_imagePath" size="75" value="<?php _e($image->imagePath); ?>" /></p>
+                    <p><strong>Image Title:</strong> <input type="text" name="edit_imageTitle" size="20" value="<?php _e($image->title); ?>" /></p>
+                    <p><strong>Image Description:</strong> <input type="text" name="edit_imageDescription" size="75" value="<?php _e($image->description); ?>" /></p>
+                    <p><strong>Sort Order:</strong> <input type="text" name="edit_imageSort" size="10" value="<?php _e($image->sortOrder); ?>" /></p>
                     <div style="clear:both;"></div>
                     <p class="major-publishing-actions left-float eg-right-margin"><input type="submit" name="Submit" class="button-primary" value="Save Image" /></p>
                     </form>                    
                     <form name="delete_image_form" action="<?php echo str_replace( '%7E', '~', $_SERVER['REQUEST_URI']); ?>" method="post">
-                    <input type="hidden" name="delete_image" value="<?php echo $image->Id; ?>" />
+                    <input type="hidden" name="delete_image" value="<?php _e($image->Id); ?>" />
+                    <?php wp_nonce_field('wpeg_delete_image', 'wpeg_delete_image'); ?>
                     <p class="major-publishing-actions left-float"><input type="submit" name="Submit" class="button-primary" value="Delete Image" /></p>
                     </form>
                 </td>
